requirements.txt
absl-py==2.2.2
annotated-types==0.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
contourpy==1.3.2
cycler==0.12.1
docker-pycreds==0.4.0
filelock==3.18.0
fonttools==4.58.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
ml_collections==1.1.0
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
platformdirs==4.3.8
protobuf==6.31.0
psutil==7.0.0
pydantic==2.11.4
pydantic_core==2.33.2
pygame==2.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sympy==1.14.0
torch==2.7.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
wandb==0.19.11